chore(shared): move and move some modules
