--- packages/shared/src/color.ts
@@ -1,8 +1,0 @@
-import color from '../compiled/picocolors';
-
-export { color };
-
-export type Colors = Omit<
-  keyof typeof color,
-  'createColor' | 'isColorSupported'
->;
