+++ packages/shared/src/index.ts
@@ -1,4 +1,3 @@
-export * from './color';
 export * from './constants';
 export * from './createHook';
 export * from './pluginStore';
@@ -8,8 +7,6 @@ export * from './getBrowserslist';
 export * from './logger';
 export * from './mergeChainedOptions';
 export * from './mergeRsbuildConfig';
-export * from './onExitProcess';
-export * from './pick';
 export * from './types';
 export * from './config';
 export * from './format';
@@ -21,7 +18,5 @@ export * from './patch';
 export * from './css';
 export * from './minimize';
 export * from './coreJs';
-export * from './progress';
-export * from './prettyTime';
 export * from './port';
 export * from './url';
