--- packages/shared/src/onExitProcess.ts
@@ -1,9 +1,0 @@
-export function onExitProcess(listener: NodeJS.ExitListener) {
-  process.on('exit', listener);
-
-  // listen to 'SIGINT' and trigger a exit
-  // 'SIGINT' from the terminal is supported on all platforms, and can usually be generated with Ctrl + C
-  process.on('SIGINT', () => {
-    process.exit(0);
-  });
-}
