+++ packages/shared/src/port.ts
@@ -1,5 +1,5 @@
 import net from 'net';
-import { color } from './color';
+import { color } from './utils';
 import { logger } from './logger';
 
 /**
@@ -72,4 +72,3 @@ export const getPort = async (
 
   return port;
 };
-/* eslint-enable no-param-reassign, @typescript-eslint/no-loop-func */
